+++ src/command.ts
@@ -1,3 +1,11 @@
+/**
+ * PLACEHOLDER
+ * 
+ * In case you're wondering this is a placeholder
+ * class. Not currently needed but here for
+ * future plans.
+ */
+
 import { KawkahCommandBase } from './base';
 import { KawkahCore } from './core';
 import { KawkahAction } from './interfaces';