+++ src/interfaces.ts
@@ -149,6 +149,7 @@ export interface IKawkahTheme {
   flag?: KawkahAnsiType;
   describe?: KawkahAnsiType;
   describeCommand?: KawkahAnsiType;
+  memo?: KawkahAnsiType;
   type?: KawkahAnsiType;
   variadic?: KawkahAnsiType;
   required?: KawkahAnsiType;
@@ -269,6 +270,7 @@ export interface IKawkahCommand {
   readonly usage?: string;
   readonly args?: string[];
   describe?: string;
+  memo?: string;
   alias?: string | string[];
   options?: IKawkahMap<string | IKawkahOption>;
   help?: string | boolean | KawkahHandler;
@@ -387,7 +389,7 @@ export interface IKawkahMiddlewareEventBase {
   completed?: number;
   elapsed?: number;
   result?: IKawkahResult;
-  isHelp?: boolean;               // indicates help was requested.
+  isActionOption?: boolean;               // indicates help was requested.
   command?: IKawkahCommandInternal;
 }